+++ packages/amplify-category-api/src/graphql-transformer/utils.ts
@@ -1,6 +1,7 @@
 import fs from 'fs-extra';
 import * as path from 'path';
-import { TransformerProjectConfig, DeploymentResources } from '@aws-amplify/graphql-transformer-core';
+import { DeploymentResources } from '@aws-amplify/graphql-transformer-interfaces';
+import { TransformerProjectConfig } from '@aws-amplify/graphql-transformer-core';
 import rimraf from 'rimraf';
 import { $TSContext, AmplifyCategories, CloudformationProviderFacade, JSONUtilities, pathManager, stateManager } from 'amplify-cli-core';
 import { CloudFormation, Fn } from 'cloudform';
@@ -204,7 +205,6 @@ export async function writeDeploymentToDisk(
   directory: string,
   rootStackFileName: string = 'rootStack.json',
   buildParameters: Object,
-  minify = false,
 ) {
   fs.ensureDirSync(directory);
   // Delete the last deployments resources except for tsconfig if present
@@ -250,7 +250,7 @@ export async function writeDeploymentToDisk(
       stackContent = JSON.parse(stackContent);
     }
     await CloudformationProviderFacade.prePushCfnTemplateModifier(context, stackContent);
-    fs.writeFileSync(fullStackPath, JSONUtilities.stringify(stackContent, { minify }));
+    fs.writeFileSync(fullStackPath, JSONUtilities.stringify(stackContent));
   }
 
   // Write any functions to disk
@@ -266,7 +266,7 @@ export async function writeDeploymentToDisk(
   }
   const rootStack = deployment.rootStack;
   const rootStackPath = path.normalize(directory + `/${rootStackFileName}`);
-  const rootStackString = minify ? JSON.stringify(rootStack) : JSON.stringify(rootStack, null, 4);
+  const rootStackString = JSON.stringify(rootStack, null, 4);
   fs.writeFileSync(rootStackPath, rootStackString);
 
   // Write params to disk