+++ packages/amplify-graphql-maps-to-transformer/src/field-mapping-lambda.ts
@@ -1,8 +1,8 @@
 import { StackManagerProvider, TransformHostProvider } from '@aws-amplify/graphql-transformer-interfaces';
-import { LambdaDataSource } from '@aws-cdk/aws-appsync';
-import * as lambda from '@aws-cdk/aws-lambda';
-import * as iam from '@aws-cdk/aws-iam';
-import * as cdk from '@aws-cdk/core';
+import { LambdaDataSource } from '@aws-cdk/aws-appsync-alpha';
+import * as lambda from 'aws-cdk-lib/aws-lambda';
+import * as iam from 'aws-cdk-lib/aws-iam';
+import * as cdk from 'aws-cdk-lib';
 import * as path from 'path';
 
 export const createMappingLambda = (host: TransformHostProvider, stackManager: StackManagerProvider) => {