+++ packages/amplify-graphql-transformer-interfaces/src/graphql-api-provider.ts
@@ -1,7 +1,10 @@
 import {
-  CfnResource, Construct, IAsset, IConstruct,
-} from '@aws-cdk/core';
-import { Grant, IGrantable, IRole } from '@aws-cdk/aws-iam';
+  CfnResource, IAsset,
+} from 'aws-cdk-lib';
+import {
+  Construct, IConstruct,
+} from 'constructs';
+import { Grant, IGrantable, IRole } from 'aws-cdk-lib/aws-iam';
 import { TransformHostProvider } from './transform-host-provider';
 
 // Auth Config Modes
@@ -110,7 +113,7 @@ export interface S3MappingFunctionCodeProvider {
 
 export type MappingTemplateProvider = InlineMappingTemplateProvider | S3MappingTemplateProvider;
 
-export interface GraphQLAPIProvider {
+export interface GraphQLAPIProvider extends IConstruct {
   readonly apiId: string;
   readonly host: TransformHostProvider;