+++ packages/amplify-schema-validator/src/__tests__/validates-fields-match-in-parent-model.test.ts
@@ -1,0 +1,15 @@
+import { validateSchema } from '..';
+import { readSchema } from './helpers/readSchema';
+
+describe('Validate Schema', () => {
+  it('fails validation when field is not in the parent model', () => {
+    const schema = readSchema('invalid-field-not-in-parent-model.graphql');
+    const errorRegex = 'email is not a field in Test';
+    expect(() => validateSchema(schema)).toThrow(errorRegex);
+  });
+
+  it('passes validation when field is in the parent model', () => {
+    const schema = readSchema('valid-field-in-parent-model.graphql');
+    expect(() => validateSchema(schema)).not.toThrow();
+  });
+});