+++ packages/amplify-dynamodb-simulator/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.6](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-dynamodb-simulator@2.4.5...amplify-category-api-dynamodb-simulator@2.4.6) (2023-03-01)
+
+**Note:** Version bump only for package amplify-category-api-dynamodb-simulator
+
+## [2.4.5](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-dynamodb-simulator@2.4.4...amplify-category-api-dynamodb-simulator@2.4.5) (2023-02-27)
+
+### Bug Fixes
+
+- **test:** update cli dependencies to use the cli rc packages ([#1294](https://github.com/aws-amplify/amplify-category-api/issues/1294)) ([7b13884](https://github.com/aws-amplify/amplify-category-api/commit/7b138841bf2c26fa16465ef263af0de7ce5a4122))
+
 ## [2.4.4](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-dynamodb-simulator@2.4.3...amplify-category-api-dynamodb-simulator@2.4.4) (2023-01-12)
 
 **Note:** Version bump only for package amplify-category-api-dynamodb-simulator