+++ packages/graphql-http-transformer/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.2.49](https://github.com/aws-amplify/amplify-category-api/compare/graphql-http-transformer@5.2.48...graphql-http-transformer@5.2.49) (2023-03-01)
+
+**Note:** Version bump only for package graphql-http-transformer
+
+## [5.2.48](https://github.com/aws-amplify/amplify-category-api/compare/graphql-http-transformer@5.2.47...graphql-http-transformer@5.2.48) (2023-02-27)
+
+**Note:** Version bump only for package graphql-http-transformer
+
+## [5.2.47](https://github.com/aws-amplify/amplify-category-api/compare/graphql-http-transformer@5.2.46...graphql-http-transformer@5.2.47) (2023-01-26)
+
+**Note:** Version bump only for package graphql-http-transformer
+
 ## [5.2.46](https://github.com/aws-amplify/amplify-category-api/compare/graphql-http-transformer@5.2.45...graphql-http-transformer@5.2.46) (2023-01-12)
 
 **Note:** Version bump only for package graphql-http-transformer