+++ packages/graphql-transformer-core/CHANGELOG.md
@@ -3,6 +3,56 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [8.0.0](https://github.com/aws-amplify/amplify-category-api/compare/graphql-transformer-core@7.6.11...graphql-transformer-core@8.0.0) (2023-03-01)
+
+**Note:** Version bump only for package graphql-transformer-core
+
+# [8.0.0-beta.6](https://github.com/aws-amplify/amplify-category-api/compare/graphql-transformer-core@8.0.0-beta.4...graphql-transformer-core@8.0.0-beta.6) (2023-02-21)
+
+**Note:** Version bump only for package graphql-transformer-core
+
+# [8.0.0-beta.5](https://github.com/aws-amplify/amplify-category-api/compare/graphql-transformer-core@8.0.0-beta.4...graphql-transformer-core@8.0.0-beta.5) (2023-02-15)
+
+**Note:** Version bump only for package graphql-transformer-core
+
+# [8.0.0-beta.4](https://github.com/aws-amplify/amplify-category-api/compare/graphql-transformer-core@7.6.10...graphql-transformer-core@8.0.0-beta.4) (2023-02-03)
+
+**Note:** Version bump only for package graphql-transformer-core
+
+# [8.0.0-beta.3](https://github.com/aws-amplify/amplify-category-api/compare/graphql-transformer-core@7.6.7...graphql-transformer-core@8.0.0-beta.3) (2022-12-27)
+
+**Note:** Version bump only for package graphql-transformer-core
+
+# [8.0.0-beta.2](https://github.com/aws-amplify/amplify-category-api/compare/graphql-transformer-core@7.6.7-beta.0...graphql-transformer-core@8.0.0-beta.2) (2022-12-12)
+
+**Note:** Version bump only for package graphql-transformer-core
+
+# [8.0.0-beta.1](https://github.com/aws-amplify/amplify-category-api/compare/graphql-transformer-core@7.6.7-beta.0...graphql-transformer-core@8.0.0-beta.1) (2022-11-30)
+
+**Note:** Version bump only for package graphql-transformer-core
+
+## [7.6.7-beta.0](https://github.com/aws-amplify/amplify-category-api/compare/graphql-transformer-core@7.6.7-cdkv2.1...graphql-transformer-core@7.6.7-beta.0) (2022-11-18)
+
+**Note:** Version bump only for package graphql-transformer-core
+
+## [7.6.7-cdkv2.1](https://github.com/aws-amplify/amplify-category-api/compare/graphql-transformer-core@7.6.7-cdkv2.0...graphql-transformer-core@7.6.7-cdkv2.1) (2022-11-15)
+
+**Note:** Version bump only for package graphql-transformer-core
+
+## [7.6.7-cdkv2.0](https://github.com/aws-amplify/amplify-category-api/compare/graphql-transformer-core@7.6.6...graphql-transformer-core@7.6.7-cdkv2.0) (2022-11-03)
+
+**Note:** Version bump only for package graphql-transformer-core
+
+## [7.6.11](https://github.com/aws-amplify/amplify-category-api/compare/graphql-transformer-core@7.6.10...graphql-transformer-core@7.6.11) (2023-02-27)
+
+### Bug Fixes
+
+- **test:** update cli dependencies to use the cli rc packages ([#1294](https://github.com/aws-amplify/amplify-category-api/issues/1294)) ([7b13884](https://github.com/aws-amplify/amplify-category-api/commit/7b138841bf2c26fa16465ef263af0de7ce5a4122))
+
+## [7.6.10](https://github.com/aws-amplify/amplify-category-api/compare/graphql-transformer-core@7.6.9...graphql-transformer-core@7.6.10) (2023-01-26)
+
+**Note:** Version bump only for package graphql-transformer-core
+
 ## [7.6.9](https://github.com/aws-amplify/amplify-category-api/compare/graphql-transformer-core@7.6.8...graphql-transformer-core@7.6.9) (2023-01-12)
 
 **Note:** Version bump only for package graphql-transformer-core